Update main.py
Alterado as entradas de inteiros para float

+++ calculadora/main.py
@@ -49,8 +49,8 @@ while( operacao != 0):
         operacaoInvalida()
         time.sleep(1.5)
         continue
-    num1 = int(input('Primeiro numero: '))
-    num2 = int(input('Digite o segundo numero: '))
+    num1 = float(input('Primeiro numero: '))
+    num2 = float(input('Digite o segundo numero: '))
     print('Resultado: ' + str(calculo(num1, num2, operacao)))
     time.sleep(1.5)
